Set up to begin implement of swap with wallets
